Remove support for v1 api
- Remove over-abstraction that is no longer needed with v1 gone (Abstract base classes for client and tests)
- Rename classes to remove "v2" as that is the only supported api
- Remove v1-only attributes from various entities
